:sparkles: Split classes into files
